Added utf-8, finally

+++ src/graphics/Graphics.h
@@ -223,8 +223,8 @@ public:
 	//
 	int drawtext(int x, int y, const char *s, int r, int g, int b, int a);
 	int drawtext(int x, int y, std::string s, int r, int g, int b, int a);
-	int drawchar(int x, int y, int c, int r, int g, int b, int a);
-	int addchar(int x, int y, int c, int r, int g, int b, int a);
+	int drawchar(int x, int y, unsigned short c, int r, int g, int b, int a);
+	int addchar(int x, int y, unsigned short c, int r, int g, int b, int a);
 
 	void xor_pixel(int x, int y);
 	void xor_line(int x, int y, int x2, int y2);
